Check for changes from old meeting listing, so cache is cleared on change

+++ includes/save.php
@@ -83,12 +83,15 @@ function tsml_save_post($post_id, $post, $update) {
 	}
 
 	// exclude_from_feeds is a checkbox
+if (!$update || empty($_POST['exclude_from_feeds']) !== empty($old_meeting->exclude_from_feeds)) {
+$changes[] = 'exclude_from_feeds';
         if (empty($_POST['exclude_from_feeds'])) {
                 delete_post_meta($post->ID, 'exclude_from_feeds');
         } else {
                 //$time_temp = $old_meeting->time;
                 update_post_meta($post->ID, 'exclude_from_feeds', $_POST['exclude_from_feeds']);
         }
+}
 
 
 	//day could be null for appointment meeting
